+++ src/components/OverlayManager/Login/types/RegisterCutomer.ts
@@ -7,12 +7,23 @@
 
 export interface RegisterCutomer_customerRegister_errors {
   __typename: "Error";
+  /**
+   * Name of a field that caused the error. A value of
+   *         `null` indicates that the error isn't associated with a particular
+   *         field.
+   */
   field: string | null;
+  /**
+   * The error message.
+   */
   message: string | null;
 }
 
 export interface RegisterCutomer_customerRegister {
   __typename: "CustomerRegister";
+  /**
+   * List of errors that occurred executing the mutation.
+   */
   errors: RegisterCutomer_customerRegister_errors[] | null;
 }