+++ src/components/PasswordResetForm/types/ResetPassword.ts
@@ -7,12 +7,23 @@
 
 export interface ResetPassword_customerPasswordReset_errors {
   __typename: "Error";
+  /**
+   * Name of a field that caused the error. A value of
+   *         `null` indicates that the error isn't associated with a particular
+   *         field.
+   */
   field: string | null;
+  /**
+   * The error message.
+   */
   message: string | null;
 }
 
 export interface ResetPassword_customerPasswordReset {
   __typename: "CustomerPasswordReset";
+  /**
+   * List of errors that occurred executing the mutation.
+   */
   errors: ResetPassword_customerPasswordReset_errors[] | null;
 }