+++ src/components/User/types/User.ts
@@ -7,6 +7,9 @@
 
 export interface User {
   __typename: "User";
+  /**
+   * The ID of the object.
+   */
   id: string;
   email: string;
   isStaff: boolean;