+++ src/views/Product/types/ProductVariantFields.ts
@@ -7,35 +7,77 @@
 
 export interface ProductVariantFields_price {
   __typename: "Money";
+  /**
+   * Currency code.
+   */
   currency: string;
+  /**
+   * Amount of money.
+   */
   amount: number;
+  /**
+   * Money formatted according to the current locale.
+   */
   localized: string;
 }
 
 export interface ProductVariantFields_attributes_attribute {
   __typename: "Attribute";
+  /**
+   * The ID of the object.
+   */
   id: string;
+  /**
+   * Name of an attribute displayed in the interface.
+   */
   name: string | null;
 }
 
 export interface ProductVariantFields_attributes_value {
   __typename: "AttributeValue";
+  /**
+   * The ID of the object.
+   */
   id: string;
+  /**
+   * Name of a value displayed in the interface.
+   */
   name: string | null;
+  /**
+   * Name of a value displayed in the interface.
+   */
   value: string | null;
 }
 
 export interface ProductVariantFields_attributes {
   __typename: "SelectedAttribute";
+  /**
+   * Name of an attribute displayed in the interface.
+   */
   attribute: ProductVariantFields_attributes_attribute;
+  /**
+   * Value of an attribute.
+   */
   value: ProductVariantFields_attributes_value;
 }
 
 export interface ProductVariantFields {
   __typename: "ProductVariant";
+  /**
+   * The ID of the object.
+   */
   id: string;
   name: string;
+  /**
+   * Quantity of a product available for sale.
+   */
   stockQuantity: number;
+  /**
+   * Price of the product variant.
+   */
   price: ProductVariantFields_price | null;
+  /**
+   * List of attributes assigned to this variant.
+   */
   attributes: ProductVariantFields_attributes[];
 }